Renamed some variables used to compute delta-nieghborhoods for columns.
Trying to figure out a good value of delta for  parm$col.delta so that the columns look reasonably similar  when they are assigned to the same neighborhoods. The trouble is most n-dimensional delta neighborhoods are singletons. When columns *are assigned to different neighborhoods, it is simply because they share the same non-zero probabilities of belonging to the different PDP clusters.