SG: The faster version of the PDP updates is in "fast_PDP.functions.R"
1. An additional argument, prob.compute.col.nbhd, is passed to function fitExample. It is the probability of recomputing the neighborhoods for the columns in an MCMC iteration.

2. There were two possible ways of updating the column allocation variables:

(i) Select a fraction (col.frac.probes) of the columns. Group them into delta neighborhoods based on similarity. Update the selected columns.

(ii) Group all parm$p columns into neighborhoods. Update a small fraction (col.frac.probes) of the neighborhoods.

I chose option (ii) because option (i) tends to produce mostly singleton neighborhoods, i.e., mostly Gibbs sampling updates. Option (ii) results in greater speedups.

In a given iteration, parm$subset_nbhd.indx has the indexes of the delta nieghborhoods that are randomly chosen to be updated in a given iteration.

3. The object posterior has the following additional components:

(a) posterior$col_exit.v: For the delta neighborhoods being updated, this is the fraction of iterations in which the MH algorithm proposed the same PDP allocation as the current allocation. In that case, we exit the MH loop without computing the MH acceptance probability.

(b) posterior$col_flip.v: This is the observed acceptance rate for the MH algorithm. The neighborhoods that exited in (a) are included in this calculation; they're counted as accepted moves.

(c) posterior$col_new_clust.v: This is the fraction of MCMC iterates for which a new cluster was occupied. It includes the case where an existing cluster was emptied and a new cluster was occupied.

(d) posterior$nbhd_max: It is computed in function "PDP_fn.check.nbhd". It is the maximum Hellinger distance between the neighborhood representative and the members, maximized over all the nieghborhoods. A relative small value indicates "tight" delta nieghborhoods. This number tends to be less than the nominal level posterior$parm$col.delta assigned in function "fn.init".